01_02_wsmt/middleware/server/requirements.txt
peewee==3.14.10
Pyro4==4.82
python-dotenv==0.20.0
